+++ test-applications/serenity-30/config/application.rb
@@ -43,5 +43,14 @@ module Serenity30
 
     # Configure sensitive parameters which will be filtered from the log file.
     config.filter_parameters += [:password]
+
+    config.aker do
+      portal :Serenity
+
+      ui_mode :custom_form
+      api_mode :http_basic
+
+      authority Aker::Authorities::Static.from_file(File.expand_path("../users.yml", __FILE__))
+    end
   end
 end