+++ CHANGELOG.md
@@ -11,6 +11,9 @@ Aker-Rails History
   renaming of the main project.
 - Switch integration test suite from Celerity to Mechanize. (#3931)
   This eliminates the JRuby dependency for integration testing.
+- Fixed: non-form UI modes can have their middleware correctly applied
+  (#1). This requires a rearrangement of your application's
+  configuration code; see the README for details.
 
 Bcsec-Rails History
 ===================